refactor(files): move policy predicate to core
